backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepagent-skills-backend"
version = "0.1.0"
description = "Terminal chat project using LangChain 1.2 and deepagents create_deep_agent."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langchain==1.2.10",
  "deepagents==0.4.4",
  "langchain-openai>=1.0.1",
  "python-dotenv>=1.1.1",
  "PyYAML>=6.0.2",
  "fastapi>=0.115.0",
  "uvicorn>=0.34.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
